✨ feat: 适配新版配置项
