+++ src/Core/Config/bot_config/base.py
@@ -32,13 +32,3 @@ class BaseModel:
     def __repr__(self):
         attrs = ", ".join(f"{key}={value!r}" for key, value in self.__dict__.items())
         return f"{self.__class__.__name__}({attrs})"
-
-
-class BaseConnection(BaseModel):
-    """通用连接配置"""
-
-    name: str = ""
-    enable: bool = False
-    token: str = ""
-    messagePostFormat: str = "array"
-    debug: bool = False
